pyproject.toml
[tool.poetry]
name = "devman-fish-shop"
version = "0.1.0"
description = ""
authors = ["Stranix <Phantom2525@gmail.com>"]
readme = "README.md"
packages = [{include = "devman_fish_shop"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pydantic = "^2.1.1"
pydantic-settings = "^2.0.2"
python-telegram-bot = "==13.*"
redis = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
